Add an extension for setting severity on comments.
This extension adds the concept of severity levels to comments. There
are three severity levels: Major, Minor, and Info. Every new comment
will have one of these levels.

The comment dialog now has new replacements for the Save button for each
severity level, which both ensures that a level is picked, and makes it
easy to save with a given level. It's far less annoying than most UIs,
which have an additional field that must be picked from.

The severity levels are then shown alongside the comment in the review
dialog, reviews, and e-mails.

This work is sponsored by NetApp.