Add a Message of the Day extension for Review Board 2.0+.
This extension allows an administrator to display a message to all
visitors to that Review Board server. This can be used to announce
incoming downtime, or a recent upgrade. It can also contain links and
other HTML formatting.

Users can dismiss the message, and it won't show up again for that
particular browser, unless cookies are cleared, or a new message is
posted.