Fix up most of the current shipped extensions.
We had a lot of silly problems with some of the current extensions.
They had less than optimal setup.py files, and were missing __init__.py
files.

rbwebhooks in particular wasn't passing required arguments to the
Extension's parent constructor.

This fixes things up so that all extensions will now load. rbxmlreview
is a bit of a special case. It's missing all the actual reviewing
functionality, and will probably need to be moved until that's there.