Add documentation for the rb-extension-pack repository.
People were a bit confused about the purpose of rb-extension-pack and it
was unclear which extensions were usable and which weren't. To fix this,
I'm adding a series of READMEs, plus a CONTRIBUTING file for GitHub, to
make it clear what the status and requirements are for each extension.

Going forward, every extension should provide a README in the same
format.