"Ship-It!" Ships ASCII Art Extension
A Review Board Extension that allows modification to the default "Ship-It!"
message when publishing a review via the "Ship It!" button.

The modifications to original "Ship-It!" text can be anything, as demonstrated
by the ship ascii art currently included by the extension. As currently
implemented, this will include the 'basic' ship, handling rich text reviews as
necessary.

There are a few TODO items, including configurability (to select other ships),
which are all listed in the 'TODO' file.

Testing Done:
Manually tested with different ascii art patterns

Reviewed at https://reviews.reviewboard.org/r/3665/