Remove the rbreports extension.
This extension is out of date, no longer works, and just confuses
people. We've had people try it, thinking they'll have detailed report
functionality, but that's not really what's in here.

Since we're working on a real reporting product, there's not much reason
to keep this maintained, so instead, it's being removed.