Add an extension for managing demo.reviewboard.org.
This provides a new authentication backend that generates suitable
usernames and passwords. Users will get a message on the login page as
to what username they should use. Upon logging in, they will be placed
in whatever configured review groups are provided on the authentication
settings page.

There's also a couple management commands for maintaining the server.
One for resetting the state of the demo server (which we can call
nightly), and one for dumping the database (similar to dumpdata, but
excludes anything we don't want to dump).