Filter stuff out from the demo database when dumping/resetting.
rbdemo now filters out a bunch of stuff we don't want to keep in the
dump file. This includes guest user entries, review request visits,
the siteconfig, and unwanted extension entries.

When loading, we preserve the old siteconfig so that we don't end up
with stuff from a development server overwriting the stuff on
production.