The Checklist Extension
The Checklist extension allows users to create a checklist when they are
reviewing a request. This is useful for cases when there are many things to
check for, and the reviewer may want to keep track of them.

The app allows:
- Auto creation of a new checklist upon viewing a new review request
- Adding new items to the checklist
- Editing existing items
- Deleting items
- Checking off items
- Undoing checked items
- Minimizing checklist
- Maximizing checklist
- Deleting checklist

The extension is a seperate Django app, and does not modify any of the
Reviewboard code. It can be enabled through the Admin page, in the Extension
tab.

Testing Done:
Manual testing on Chrome and Firefox:
- Created checklists by viewing new review requests
- Added items on the checklist
- Deleted items on checklist
- Modified items on checklist
- Checking off items
- Undoing checked items
- Tested on different file attachments but same requests
- Tested on different review requests
- Tested on different users

This commit does not include the images which were used in the original change,
because they weren't posted (and I don't see a fork that includes them). The
missing files are:
- checklist/checklist/static/images/cross.png
- checklist/checklist/static/images/exit.png
- checklist/checklist/static/images/max.png
- checklist/checklist/static/images/min.png
- checklist/checklist/static/images/plus.png

I'm also not sure why there's both a "style.less" and "index.css" file. It
looks like one is derived from the other, but it's the .css file which is
bundled in the extension. We'll have to investigate this.

Reviewed at https://reviews.reviewboard.org/r/5007/