[rbdemo] Create media/images/ after resetting media files.
reset-demo was wiping the media directory and copying over the new
files, but these didn't necessarily have an images/ directory. This
meant that the directory wouldn't be created, and the Manual Updates
would kick in.

We now explicitly create the required directories if they don't exist.

Reviewed at https://reviews.reviewboard.org/r/5877/