Modernize the packaging for rbmotd and rbseverity.
rbmotd and rbseverity now define their version information in a set of
functions and variables in their top-level packages.

setup.py has been updated to use these, and to also remove the
dependency information we don't want to require there (as it can mess up
an install if not running a new enough version of Review Board).

Classifiers have been added to help organize these packages on PyPI.

There's also a .beanbag-build.cfg for each, which are used to make it
easy to build and release these packages.

Reviewed at https://reviews.reviewboard.org/r/6025/