Add an extension for a "Note to Reviewers" field on review requests.
This extension adds a "Note to Reviewers" field below "Testing Done,"
which developers can use to give instructions on what to focus on or
to pre-answer questions that may be asked, without lumping those into
the description of the change.

This field is saved and published along with the draft.

While useful, it largely functions as a demo for how to add a simple
field.

Reviewed at https://reviews.reviewboard.org/r/6026/