Changed style of Checklist extension to fit RB colour scheme.
This change makes several alterations to the style for the checklist. The
checklist is now drawn in a light blue, matching the colors used in the rest of
Review Board. This also eliminates the icons in favor of text links.

Testing done:
Ran with the checklist extension enabled and saw the new styles.

Reviewed at https://reviews.reviewboard.org/r/5939/