[rbnotefield] Work around a breakage with rendering change entries.
Review Board < 2.0.7 breaks when trying to render changes to the Note
field if adding content for the first time after the review request has
already been published. The old value for the field was 'None', which
broke the rendering due to the expectation of strings.

While Review Board 2.0.7 is getting a fix for this, we still want to
work around it for earlier releases. This adds a new
render_change_entry_html() function that just normalizes the values
before calling the parent function, ensuring we're dealing with empty
strings.

Fixes bug #3549
Reviewed at https://reviews.reviewboard.org/r/6291/