[rbslack] Fix a crash when review requests are closed/reopened by hooks.
When closing a review request through a post-commit hook, or any other
means where a user isn't provided, rbslack would crash. Same with
reopening without a provided user. Now, if the user is None, it will
default to the owner of the review request.

The other signals all have a valid user passed in, so they don't need
this sort of check.

Reviewed at https://reviews.reviewboard.org/r/6296/