[review together] Allow the user to specify their own hub in the extension config.
With TogetherJS, all communication is done via WebRTC, but people may not be
pleased having the negotiation happen via Mozilla's public hub. This change
adds configuration to allow people to specify their own hub.

Testing Done:
A hub server was run locally. Its URL was used in the extension configuration.
The local server was used successfully to support the extension. To test the
hub used is actually the one indicated, tried to launch the chat app while the
local server is not on. The chat would still launch but using the "add friend"
link on another browser could not connect the two clients.

To test the default hub is used when no hub is specified, remove any input in
the extension configuration screen. Launch the chat app and using another
browser connect to the first client with the "Add friend" link. The clients
should be able to see each other's mouse movements.

Reviewed at https://reviews.reviewboard.org/r/6102/