[rbdemo] Add some checks to prevent us from breaking the demo site.
This adds some pre-reset checks to rbdemo's reset-demo command that
ensures we have all the proper permissions and abilities we need prior
to resetting the demo database state. This includes read-access to the
fixtures and media directories, write access to the destination media
directory, and the ability to change ownership. It also validates that
the specified user and group are valid.

Reviewed at https://reviews.reviewboard.org/r/6338/