[review-together] Display horizontal bar which indicates peer's scroll position.
This change adds a horizontal bar for each peer, showing where on the page
they've scrolled. This is a good prototype for some kind of google docs or
hackpad-like feature, where we could show an avatar of some sort and be able to
jump to each others' locations.

Testing done:
Launched TogetherJS in one browser. Opened a second browser and entered the URL
created by the first browser. Scrolled around on both browsers to check if
horizontal bar displayed and moved with the other browser.

Demostration Video:
http://youtu.be/OO7Qc_9jiz8

Reviewed at https://reviews.reviewboard.org/r/6194/