[review-together] Cleanups to scroll position work
Peter Tran did a lot of preparatory work for showing the scroll
position of your peers in a TogetherJS session. This work was
landed in a project branch, awaiting some clean-up once the term
was over.

This is the clean-up I had in mind. I've fixed the following:

* Scroll position indicators now have a high z-index to avoid being
  hidden behind other elements.
* If no hub URL has been specified, properly falls back to the Mozilla
  TogetherJS hub.
* Moves the ReviewTogether JS code into a closure to avoid polluting the
  global namespace.
* If a peer's colour changes during a session, this is now manifested
  for everybody in the session without requiring a page reload.
* Switched a lot of double-quotes to single-quotes.
* Leaving a TogetherJS session removes all scroll position indicators.
* When somebody leaves a session, their scroll position indicator is
  removed.

Reviewed at https://reviews.reviewboard.org/r/6884/