[rbdemo] Delete the SCMTools prior to loading in fixtures.
It was easy to get into a situation where SCMTools from the fixtures
conflicted with what was populated from registerscmtools. We now clear
out the SCMTools in the database immediately before we load the fixture
data.

Testing Done:
This fixed a problem I had on demo.reviewboard.org with deploying updates
to the fixtures.

Reviewed at https://reviews.reviewboard.org/r/7138/