Add a review stopwatch extension.
This adds a new extension, the review stopwatch. The goal here is for reviewers
to be able to track the amount of time that they spend reviewing a change, and
have that information available in the API.

The stopwatch UI is shown in the lower-right corner of the page. Clicking it
will toggle on and off the timer. The resulting total seconds spent reviewing
is stored in the review extra_data field.

The one somewhat confusing thing here is that if a user opens up multiple tabs
related to the same review request (say, review request details, diff viewer,
and a couple file attachments), there will be timers on each page. If these are
run independently, they'll step on each other. We'll just have to document how
this should be used and hope that it doesn't become an issue.

Testing done:
- Tested that toggle on/off worked correctly.
- Tested that the current time was saved to the review's extra_data when
  toggling off.
- Tested display of the time in the stopwatch overlay, review dialog, and
  review boxes for various different values.
- Tested that the display in the review dialog updated when the stopwatch was
  still running.
- Tested that publishing or discarding the review stopped the stopwatch
  correctly, and (in the case of publish), saved correctly.
- Tested onbeforeunload handler.

Reviewed at https://reviews.reviewboard.org/r/7429/