Add a simple "user stats" extension that uses the new infobox features.
This change adds a new extension that populates the extra section of the user
infobox with counts of that user's incoming and outgoing review requests.

Testing done:
Installed and enabled the extension. Hovered over a user and saw counts of
their incoming and outgoing review requests.

Reviewed at https://reviews.reviewboard.org/r/8080/