[checklist] Move checklist view instantiation into pure JS.
The way that the checklist view was created and added to the page was pretty
out of date. I've updated it to use the `PageManager` to wait until the page is
ready and then fetch the review request ID from that. The view is then created
and added to the page like we do with pretty much everything else, instead of
creating a dummy element and having the view add itself.

Testing done:
Loaded the page and saw the checklist was visible.

Reviewed at https://reviews.reviewboard.org/r/8621/