[checklist] Clean up the markup, style, and editing for checklists.
The markup and CSS for the checklist was pretty ugly, and the way that items
were edited was awkward (cramming an `<input>` element into the row and
requiring users to know to hit "enter" or "escape"). This change cleans things
up in the implementation, polishes up the look a bit, and changes editing to
use `contenteditable` on the text element.

Testing done:
Created, edited, checked/unchecked, and removed items.

Reviewed at https://reviews.reviewboard.org/r/8651/