[rbseverity] Fix the severity extension to properly support general comments
The severity of a problem was not being displayed on the review page, if the
severity was added using the add comment button on the review dialog page.

The drop downs menus for severity options on the edit review page were not
saving the model with the selected option. The model not saving the selected
option caused the data not be saved in the database when, submitting the form.

The patch fixes the problem by saving the model anytime the severity changes in
the drop down. When the severity is saved it is properly added to the database,
and this fixes the problem on the review page.

Testing Done:
I added a comment of each severity using the add comment button on the
reviewable page. On the review dialog page I added another comment with a
severity and published it from this page. I repeated the process adding
multiple comments with varying degrees of severity, and different amounts of
comments.

I ran the js tests with the extension enabled and all expected tests passed. I
had 2 failures, they seemed unrelated and other people were experiencing them,
and they are being filed separately.

All tickets had proper severity in the database and displayed it on the
reviewable page.

Bugs closed: 4428

Reviewed at https://reviews.reviewboard.org/r/9205/