[rbslack] Modernize rbslack and add a deprecation warning.
This change modernizes the rbslack extension to work with Review Board
4.0 and Django 1.11. This involves updating the urlpatterns and the
handler for the review_request_closed signal, and fixing up
documentation throughout.

I've also decided to add a warning to the extension's description that
it is deprecated and will be removed soon, so hopefully all this work is
actually moot.

Testing Done:
Loaded the slack integration and saw that things worked as expected.

Reviewed at https://reviews.reviewboard.org/r/11058/