[rbxmlreview] Modernize the XML Review extension.
The XML review UI extension was pretty long in the tooth, and wasn't
using static media or the review UI JavaScript infrastructure in the
correct way. This change updates it so that it doesn't need its own
template, and relies on static media bundles instead of manually loading
the JavaScript file.

Testing Done:
Enabled the extension. Uploaded an XML file and was able to review it.

Reviewed at https://reviews.reviewboard.org/r/11081/