[rbcommenttype] Update to be compatible with Review Board 4 and Python 3.
This fixes some string, import, and URL building issues that prevented
it from installing and working on Review Board 4 and Python 3. These
were pretty minor issues that got missed when updating other extensions
in the extension pack.

Testing Done:
Installed the extension on Python 3 and enabled it. Successfully
configured it with old and new comment types.

Saw the resulting options when making a comment, in the review dialog,
and in published comments.

Reviewed at https://reviews.reviewboard.org/r/11709/