Delete rbslack
Once upon a time, integration with Slack was done through an extension
that we shipped. This wasn't a great implementation and was quite
limited, and we ended up completely rewriting it for the
`rbintegrations`, which is now always installed alongside Review Board.
We kept the old `rbslack` implementation in git, but Review Board has
now included integrations for a long time, and this codebase is no
longer useful. This change gets rid of it.

Reviewed at https://reviews.reviewboard.org/r/12181/