Fix and update the rbchecklist extension for Review Board 5.0.
This change updates the rbchecklist extension for Review Board 5.0.x and
Django 3.2. This extension was also broken in a few ways, and I've fixed
up the things that were causing it to break (some problems with error
handling on the JS side, and a missing HTML element that was causing the
settings page to show poorly).

I'd really like to start over on this extension--what we have is both
under-featured and over-complex, and the visual design is highly
lacking. For now, this change gets things limping along again.

Testing Done:
Installed and activated the extension. Was able to use the features
successfully.

Reviewed at https://reviews.reviewboard.org/r/12202/