Port rbmotd to Review Board 7 and Ink, and improve accessibility.
This is a long-overdue change to make rbmotd work with Review Board 7
(and, more specifically, Django 4.2). This primarily deals with import
updates.

We now use Ink styling for the message banner, ensuring light and dark
mode work correctly. The close button has been moved to the right-hand
side of the banner, and accessibility information has been added for all
parts of the banner.

Testing Done:
Tested on light and dark modes.

Checked the accessibility tree for the right information.

Reviewed at https://reviews.reviewboard.org/r/14357/