Prepare rbmotd 3 for release.
This adds a new README for rbmotd and updates the package information
for the new requirements for Review Board 7.

We ideally would be using pyproject.toml, but the support for
pyproject-based extension packages depends on the in-development
Review Board 7.1.

Testing Done:
Checked the README rendering. Tested links and spelling.

Tested building new packages.

Reviewed at https://reviews.reviewboard.org/r/14358/