[rbnotefield] Update for modern Django.
rbnotefield is a trivially small extension to add a "Note to reviewers"
text field. This extension was never updated for modern Django.

The only incompatibility was the use of `ugettext_lazy`, which now needs
to be `gettext_lazy`. Since it's so tiny, while I was here I added in
type hints.

Testing Done:
Installed and activated the extension.

Reviewed at https://reviews.reviewboard.org/r/14451/